Migrate away from deprecated componentWillReceiveProps
This implementation is not ideal, but it works. The preferred alternative, getDerivedStateFromProps, does not have access to the previous properties and thus is insufficient for comparing new nodes to the old ones. This comparison is needed to determine whether the expensive node flattening method should be called or skipped.

Alternatively, the previous props could be stored in the state and thus be recalled in getDerivedStateFromProps.